+++ intercom_flutter/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## 9.6.17
+
+* Bump Intercom Android SDK version to [18.5.0](https://github.com/intercom/intercom-android/releases/tag/18.5.0)
+* Bump Intercom iOS SDK version to [19.7.0](https://github.com/intercom/intercom-ios/releases/tag/19.7.0)
+* Added API `suppressProactiveContent` to control which proactive content types (carousels, surveys) can be displayed. Not supported on web.
+* Updated dependency `intercom_flutter_platform_interface: ^2.0.8`
+* Updated dependency `intercom_flutter_web: ^1.1.13`
+
+## 9.6.16
+
+* Bump Intercom Android SDK version to [18.4.0](https://github.com/intercom/intercom-android/releases/tag/18.4.0)
+
+## 9.6.15
+
+* Bump Intercom iOS SDK version to [19.6.5](https://github.com/intercom/intercom-ios/releases/tag/19.6.5)
+
+## 9.6.14
+
+* Bump Intercom Android SDK version to [18.3.2](https://github.com/intercom/intercom-android/releases/tag/18.3.2)
+* Bump Intercom iOS SDK version to [19.6.4](https://github.com/intercom/intercom-ios/releases/tag/19.6.4)
+
+## 9.6.13
+
+* Bump Intercom Android SDK version to [18.3.1](https://github.com/intercom/intercom-android/releases/tag/18.3.1)
+
+## 9.6.12
+
+* Bump Intercom Android SDK version to [18.3.0](https://github.com/intercom/intercom-android/releases/tag/18.3.0)
+
+## 9.6.11
+
+* Bump Intercom iOS SDK version to [19.6.3](https://github.com/intercom/intercom-ios/releases/tag/19.6.3)
+
+## 9.6.10
+
+* Bump Intercom iOS SDK version to [19.6.2](https://github.com/intercom/intercom-ios/releases/tag/19.6.2)
+
 ## 9.6.9
 
 * Bump Intercom iOS SDK version to [19.6.1](https://github.com/intercom/intercom-ios/releases/tag/19.6.1)